requirements.txt
ultralytics==8.1.0
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
Pillow==10.2.0
PyYAML==6.0.1
matplotlib==3.8.2
